Refactor Session Management and Introduce Audio Sink
- Updated the Session class to use a new initialization method, removing the connect method and replacing it with a context manager approach.
- Introduced a new SessionConnectError exception to handle connection failures more gracefully.
- Added an AudioSink class to manage audio playback from the session, ensuring audio frames are processed correctly.
- Modified the README to reflect changes in the Session usage and frame handling.
- Updated various examples to align with the new Session initialization and audio handling.
- Refactored frame handling to utilize event iterators for better performance and responsiveness.
- Adjusted FPS calculation in the FpsThread to remove unnecessary pull time tracking.
- Enhanced Vulkan and CUDA frame handling to ensure compatibility with the new Session structure.
- Improved audio queue management in the audio handler to prevent overflow and ensure smooth playback.

+++ pybind/src/audio_handler.cpp
@@ -27,7 +27,7 @@ void AudioHandler::QueueFrame(int16_t *buf, size_t samples_count)
     // If nobody drains the queue, drop the oldest samples to cap latency and memory.
     size_t max_samples = audio_queue_max_frames
                              ? audio_queue_max_frames * audio_channels
-                             : 3 * audio_buffer_size / sizeof(int16_t); // audio_buffer_size is in bytes, same threshold as chiaki-ng
+                             : 3 * (size_t)audio_buffer_size / sizeof(int16_t); // audio_buffer_size is in bytes, same threshold as chiaki-ng
     if (audio_queue.size() > max_samples)
     {
         if (!audio_queue_overflow_logged)
@@ -78,7 +78,7 @@ size_t AudioHandler::GetAudioQueueMaxFrames()
     std::lock_guard<std::mutex> lock(audio_buf_mutex);
     if (audio_queue_max_frames)
         return audio_queue_max_frames;
-    return audio_out_sample_size ? 3 * audio_buffer_size / audio_out_sample_size : 0;
+    return audio_out_sample_size ? 3 * (size_t)audio_buffer_size / audio_out_sample_size : 0;
 }
 
 void AudioHandler::SetAudioQueueMaxFrames(size_t max_frames)
